requirements.txt
streamlit==1.35.0
pandas==2.2.2
fpdf2
markdown2
openai
python-dotenv
PyPDF2
pytesseract
Pillow
pinecone==7.3.0
sentence-transformers==5.0.0